requirements.txt
Flask
flask-cors
psycopg2-binary
python-dotenv
google-genai
forex-python==1.16
